bot/requirements.txt
boto3==1.22.7
backoff==2.0.1
python-telegram-bot==20.0a0